Finally got the program working
It took ages to realize that the problem was not, in fact, my code, but my board. I finally used someone else's board with my old code just to double check, and saw that the past 96 hours of programming were spent on fixing an issue with my FPGA board, which is non-fixable with code.